pyproject.toml
[project]
name = "llm_connector"
version = "0.1.0"
description = "A simple Python utility to connect with multiple LLM providers"
authors = [{ name = "Your Name", email = "your.email@example.com" }]
dependencies = [
    "langchain-core>=0.1.27",
    "langchain-openai>=0.0.8",
    "langchain-anthropic>=0.1.4",
    "langchain-google-genai>=0.0.11",
    "python-dotenv>=1.0.1",
    "langchain-ollama>=0.1.6"
]
requires-python = ">=3.9"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true 
